Using bindgen data everywhere

+++ darwin-libproc/src/name.rs
@@ -4,8 +4,9 @@ use std::os::unix::ffi::OsStringExt;
 
 /// Fetch process name for `pid` provided.
 pub fn name(pid: libc::pid_t) -> io::Result<OsString> {
-    let mut buffer: Vec<u8> =
-        Vec::with_capacity(darwin_libproc_sys::PROC_PIDPATHINFO_MAXSIZE);
+    let mut buffer: Vec<u8> = Vec::with_capacity(
+        darwin_libproc_sys::PROC_PIDPATHINFO_MAXSIZE as usize,
+    );
 
     let result = unsafe {
         darwin_libproc_sys::proc_name(
